% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.train.R
\name{auto.train}
\alias{auto.train}
\title{auto.train}
\usage{
auto.train(x, y, model, trControl)
}
\arguments{
\item{x}{the estimation dataset, a matrix X of covariate observations used for the estimation step}

\item{y}{a vecteur Y of the target observation used for the estimation step}

\item{model}{a character string, name of the model to compute}

\item{trControl}{control parameter for the training of the models, see \code{\link{caret:train}}}
}
\description{
automatic training of a model with error handling
}
\examples{
y<-rnorm(10)
ychap<-rep(0,10)
rmse(y,ychap,digits=4)
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}

