% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIndicator.R
\name{createIndicators}
\alias{createIndicators}
\title{createIndicators}
\usage{
createIndicators(Data, factor.name, prefix = NULL)
}
\arguments{
\item{Data}{the dataset, containing observations of covariates}

\item{factor.name}{name of the factor covariate}

\item{prefix}{a character string to prefix the names of indicators covariates}
}
\description{
convert a factor covariate in a data.frame to a set of indictors covariates
}
\examples{
to complete
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}

