% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPast.R
\name{createPast}
\alias{createPast}
\title{createPast}
\usage{
createPast(Data, name, past = 1:24)
}
\arguments{
\item{Data}{the dataset, containing observations of covariates}

\item{name}{name of the covariate}

\item{past}{a vecteur of lags}
}
\description{
create a new covariate called name.i, i is the lag, which is a lagged version of the covariate name
}
\examples{
to complete
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}

