% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_1_data.R
\name{get_level_1_data}
\alias{get_level_1_data}
\title{Get level 1 training and testing data}
\usage{
get_level_1_data(training_frame, testing_frame, response, model_wrappers,
  n_folds = 5)
}
\arguments{
\item{response}{string, name of the response column}

\item{model_wrappers}{named vector of functions, names are optional, and
in their absence names will be automatically generated.  See details below
for defining model wrappers.}

\item{n_folds}{}
}
\description{
Get level 1 training and testing data
}

