% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error}
\usage{
mape(y, ychap, digits = 3)
}
\arguments{
\item{y}{the observations to be predicted}

\item{ychap}{the predictions}

\item{digits}{the precision in number of digits}
}
\value{
MAPE in percentage
}
\description{
compute the Mean Absolute Percentage Error of a prediction vector
}
\examples{
y<-rnorm(10)
ychap<-rep(0,10)
mape(y,ychap,digits=4)
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}

