% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(y, ychap, digits = 3)
}
\arguments{
\item{y}{the observations to be predicted}

\item{ychap}{the predictions}

\item{digits}{the precision in number of digits}
}
\value{
a positive real number the RMSE
}
\description{
compute the Root Mean Square Error of a prediction vector
}
\examples{
y<-rnorm(10)
ychap<-rep(0,10)
rmse(y,ychap,digits=4)
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}

