% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollArima.R
\name{rollArima}
\alias{rollArima}
\title{rollArima}
\usage{
rollArima(arima.model, ynew, horizon = 1)
}
\arguments{
\item{arima.model}{an arima model, from forecast::auto.arima()}

\item{ynew}{the time serie to predict}

\item{horizon}{the forecast horizon}
}
\description{
rolling forecasts for ARIMA models
}
\examples{
y<-rnorm(10)
ychap<-rep(0,10)
rmse(y,ychap,digits=4)
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}

