% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollETS.R
\name{rollETS}
\alias{rollETS}
\title{rollETS}
\usage{
rollETS(ets.model, ynew, horizon = 1)
}
\arguments{
\item{ets.model}{an exponential smoothing model, from forecast::ets()}

\item{ynew}{the time serie to predict}

\item{horizon}{the forecast horizon}
}
\description{
rolling forecasts for exponential smoothing models (forecast::ets)
}
\examples{
y<-rnorm(10)
ychap<-rep(0,10)
rmse(y,ychap,digits=4)
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}

