% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollHW.R
\name{rollHW}
\alias{rollHW}
\title{rollHW}
\usage{
rollHW(hw.model, ynew, horizon = 1)
}
\arguments{
\item{hw.model}{a Holt Winters exponential smoothing model, from forecast::HoltWinters()}

\item{ynew}{the time serie to predict}

\item{horizon}{the forecast horizon}
}
\description{
rolling forecasts for Holt Winters exponential smoothing models (forecast::HoltWinters)
}
\examples{
y<-rnorm(10)
ychap<-rep(0,10)
rmse(y,ychap,digits=4)
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}

