% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrmse.R
\name{rrmse}
\alias{rrmse}
\title{Relative Root Mean Square Error}
\usage{
rrmse(y, ychap, digits = 3)
}
\arguments{
\item{y}{the observations to be predicted}

\item{ychap}{the predictions}

\item{digits}{the precision in number of digits}
}
\value{
a positive real number the RRMSE
}
\description{
compute the Relative Root Mean Square Error of a prediction vector
}
\examples{
y<-rnorm(10,5,1)
ychap<-rep(0,10)
rrmse(y,ychap,digits=4)
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}

