% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking.r
\name{stacking}
\alias{stacking}
\title{stacking}
\usage{
stacking(x0, x1, y0, train.rate, model.list, trControl, data.export = FALSE,
  sample.type = "random", block.size = NULL)
}
\arguments{
\item{x0}{the estimation dataset, a matrix X of covariate observations used for the estimation step}

\item{x1}{the prediction dataset, a matrix X of covariate observations used for the prediction step}

\item{y0}{a vecteur Y of the target observation used for the estimation step}

\item{train.rate}{a real number in [0,1] corresponding to the rate of observation used for model computation, randomly picked (how depends on sample.type)
n1=floor(train.rate*n) observations are used for model computation, n-n1 observations for their stacked versions}

\item{model.list}{a list of model to compute, already testes ones are:
projection based method: \code{projectionList<-c("pcr","ppr","pls","plsRglm","simpls","spls")}
trees: \code{treeList<-c("cubist","gbm","blackboost","ctree","ctree2","rpart1SE","rpart2","treebag","xgbTree")}
additive models: \code{additiveList<-c("bagEarth","bagEarthGCV", "bstTree","earth","gamLoess","gamSpline","gcvEarth")}
kernel methods: \code{kernelList<-c("kernelpls","kknn","svmLinear","svmPoly","svmRadial","svmRadialSigma","svmRadialCost","knn","kknn")}}

\item{trControl}{control parameter for the training of the models, see \code{\link[caret]{train}}}

\item{data.export}{a boolean indicating wether or not the datasets (including models' prediction) should be returned or not}

\item{sample.type}{the type of sampling method to split the observations in the dataset into two, one for model estimation, one for stacking. Could be: "random":
random sampling without replacement, "consecutive": just the first n1 rows or "blocks": data are picked by blocks of size \code{block.size} could be useful 
for time series}
}
\value{
a list containing:
\item{model.list}{a list of models' output, as many elements as model in the argument model.list}
\item{forecast}{the prediction of the models on the prediction dataset}
\item{forecast.stack}{the prediction of the stacked models on the prediction dataset}
\item{Xnew}{a matrix X of covariate observations used for the stacked estimation step, return only when \code{data.export=TRUE}}
\item{Ynew}{a vector Y of covariate observations used for the stacked estimation step, return only when \code{data.export=TRUE}}
}
\description{
compute a list of model and their stacked versions
}
\examples{
y<-rnorm(10)
ychap<-rep(0,10)
rmse(y,ychap,digits=4)
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}

